code refactor: extracting class, updating copyright, adding "namespace" to settings class.
